tourism_agent/tourism_agent_fixed/requirements.txt
# ── AI / LLM ──────────────────────────────────────────────
google-generativeai>=0.7.0

# ── Trend discovery ───────────────────────────────────────
pytrends>=4.9.2
praw>=7.7.1
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
duckduckgo-search>=6.1.0
newsapi-python>=0.2.7

# ── Image processing ──────────────────────────────────────
Pillow>=10.0.0
requests>=2.31.0

# ── Video creation ────────────────────────────────────────
moviepy>=1.0.3
numpy>=1.24.0
ffmpeg-python>=0.2.0

# ── Audio processing ──────────────────────────────────────
gTTS>=2.4.0
pydub>=0.25.1
scipy>=1.11.0
elevenlabs>=1.3.0

# ── Social media publishing ───────────────────────────────
python-telegram-bot>=20.7

# ── Scheduling ────────────────────────────────────────────
APScheduler>=3.10.4

# ── Utilities ─────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.2
aiohttp>=3.9.0
colorama>=0.4.6
